[hdf5] Fixed a compilation bug for HDF5 detected on Windows XP GCC 492
[ 15%] Building CXX object interface_hdf5/CMakeFiles/interface_hdf5.dir/src/h5_tools.cpp.obj

c:/Jenkins_slave/workspace/XDyn_Release_XP32_GCC492_PCGJ_130_66_124_210/code/interface_hdf5/src/h5_tools.cpp: In function 'void H5_Tools::read(const H5::H5File&, const string&, std::vector<std::vector<double> >&)':
c:/Jenkins_slave/workspace/XDyn_Release_XP32_GCC492_PCGJ_130_66_124_210/code/interface_hdf5/src/h5_tools.cpp:584:75: error: conversion to 'size_t {aka unsigned int}' from 'hsize_t {aka long long unsigned int}' may alter its value [-Werror=conversion]
     double *p = (double*)malloc(dims_out[0] * dims_out[1] * sizeof(double));
                                                                           ^
compilation terminated due to -Wfatal-errors.
cc1plus.exe: all warnings being treated as errors

+++ code/interface_hdf5/src/h5_tools.cpp
@@ -581,7 +581,7 @@ void H5_Tools::read(
         THROW(__PRETTY_FUNCTION__, InternalErrorException, "One expects 2 rank data, not" << ndims);
     }
     H5::DataSpace memspace(2, dims_out);
-    double *p = (double*)malloc(dims_out[0] * dims_out[1] * sizeof(double));
+    double *p = (double*)malloc((size_t) (dims_out[0] * dims_out[1]) * sizeof(double));
     dataset.read(p, H5::PredType::NATIVE_DOUBLE, memspace, dataspace);
     for (hsize_t i=0; i<dims_out[0];++i)
     {
@@ -628,7 +628,7 @@ void H5_Tools::read(
         THROW(__PRETTY_FUNCTION__, InternalErrorException, "One expects 3 rank data, not" << ndims);
     }
     H5::DataSpace memspace(3, dims_out);
-    double *p = (double*)malloc(dims_out[0] * dims_out[1] * dims_out[2] * sizeof(double));
+    double *p = (double*)malloc((size_t) (dims_out[0] * dims_out[1] * dims_out[2]) * sizeof(double));
     dataset.read(p, H5::PredType::NATIVE_DOUBLE, memspace, dataspace);
     for (hsize_t i=0; i<dims_out[0];++i)
     {
